Fixed Audio
I fixed the audio, basicaly when you called play, it was returning something other then the object you just created. Just dont call play when you load audio and it should be fine. Also, this is a slightly older version of HowlerJS, you should be able to update it to the current version if you want, its just the one I had on my computer that I am used to.